# Add Hardware Accelerated Checksums