--- a/tests/Directory.Build.targets
+++ b/tests/Directory.Build.targets
@@ -34,6 +34,7 @@
     <PackageReference Update="Microsoft.NET.Test.Sdk" Version="16.5.0-preview-20200116-01" />
     <PackageReference Update="Moq" Version="4.10.0" />
     <PackageReference Update="Pfim" Version="0.9.1" />
+    <PackageReference Update="SharpZipLib" Version="1.2.0" />
     <PackageReference Update="System.Drawing.Common" Version="4.7.0" />
     <PackageReference Update="xunit" Version="2.4.1" />
     <PackageReference Update="xunit.runner.visualstudio" Version="2.4.1" />