… src/ImageSharp/Formats/Png/Zlib/Crc32.Lut.cs
… src/ImageSharp/Formats/Png/Zlib/IChecksum.cs
… src/ImageSharp/Formats/Png/Zlib/README.md
… src/ImageSharp/Formats/Png/Zlib/fast-crc-computation-generic-polynomials-pclmulqdq-paper.pdf
… tests/ImageSharp.Benchmarks/General/Adler32Benchmark.cs
… tests/ImageSharp.Benchmarks/General/Crc32Benchmark.cs
… tests/ImageSharp.Tests/Formats/Png/Adler32Tests.cs
… tests/ImageSharp.Tests/Formats/Png/Crc32Tests.cs
… tests/ImageSharp.Tests/Formats/Png/PngDecoderTests.Chunks.cs