### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp/pulls) open
- [x] I have verified that I am following matches the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description
<!-- A description of the changes proposed in the pull-request -->
Added support for hardware accelerated Crc32 and Adler32 checksum generation both used by our png codecs. They're fast!

Benchmarks.

**Adler32**

|               Method |       Runtime | Count |        Mean |       Error |    StdDev | Ratio | RatioSD | Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------------- |-------------- |------ |------------:|------------:|----------:|------:|--------:|------:|------:|------:|----------:|
| SharpZipLibCalculate |    .NET 4.7.2 |  1024 |   793.18 ns |   775.66 ns | 42.516 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|   SixLaborsCalculate |    .NET 4.7.2 |  1024 |   384.86 ns |    15.64 ns |  0.857 ns |  0.49 |    0.03 |     - |     - |     - |         - |
|                      |               |       |             |             |           |       |         |       |       |       |           |
| SharpZipLibCalculate | .NET Core 2.1 |  1024 |   790.31 ns |   353.34 ns | 19.368 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|   SixLaborsCalculate | .NET Core 2.1 |  1024 |   465.28 ns |   652.41 ns | 35.761 ns |  0.59 |    0.03 |     - |     - |     - |         - |
|                      |               |       |             |             |           |       |         |       |       |       |           |
| SharpZipLibCalculate | .NET Core 3.1 |  1024 |   877.25 ns |    97.89 ns |  5.365 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|   SixLaborsCalculate | .NET Core 3.1 |  1024 |    45.60 ns |    13.28 ns |  0.728 ns |  0.05 |    0.00 |     - |     - |     - |         - |
|                      |               |       |             |             |           |       |         |       |       |       |           |
| SharpZipLibCalculate |    .NET 4.7.2 |  2048 | 1,537.04 ns |   428.44 ns | 23.484 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|   SixLaborsCalculate |    .NET 4.7.2 |  2048 |   849.76 ns | 1,066.34 ns | 58.450 ns |  0.55 |    0.04 |     - |     - |     - |         - |
|                      |               |       |             |             |           |       |         |       |       |       |           |
| SharpZipLibCalculate | .NET Core 2.1 |  2048 | 1,616.97 ns |   276.70 ns | 15.167 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|   SixLaborsCalculate | .NET Core 2.1 |  2048 |   790.77 ns |   691.71 ns | 37.915 ns |  0.49 |    0.03 |     - |     - |     - |         - |
|                      |               |       |             |             |           |       |         |       |       |       |           |
| SharpZipLibCalculate | .NET Core 3.1 |  2048 | 1,735.11 ns | 1,374.22 ns | 75.325 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|   SixLaborsCalculate | .NET Core 3.1 |  2048 |    87.80 ns |    56.84 ns |  3.116 ns |  0.05 |    0.00 |     - |     - |     - |         - |
|                      |               |       |             |             |           |       |         |       |       |       |           |
| SharpZipLibCalculate |    .NET 4.7.2 |  4096 | 3,054.53 ns |   796.41 ns | 43.654 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|   SixLaborsCalculate |    .NET 4.7.2 |  4096 | 1,538.90 ns |   487.02 ns | 26.695 ns |  0.50 |    0.01 |     - |     - |     - |         - |
|                      |               |       |             |             |           |       |         |       |       |       |           |
| SharpZipLibCalculate | .NET Core 2.1 |  4096 | 3,223.48 ns |    32.32 ns |  1.771 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|   SixLaborsCalculate | .NET Core 2.1 |  4096 | 1,547.60 ns |   309.72 ns | 16.977 ns |  0.48 |    0.01 |     - |     - |     - |         - |
|                      |               |       |             |             |           |       |         |       |       |       |           |
| SharpZipLibCalculate | .NET Core 3.1 |  4096 | 3,672.33 ns | 1,095.81 ns | 60.065 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|   SixLaborsCalculate | .NET Core 3.1 |  4096 |   159.44 ns |    36.31 ns |  1.990 ns |  0.04 |    0.00 |     - |     - |     - |         - |

**Crc32**

|               Method |       Runtime | Count |         Mean |        Error |       StdDev | Ratio | RatioSD | Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------------- |-------------- |------ |-------------:|-------------:|-------------:|------:|--------:|------:|------:|------:|----------:|
| SharpZipLibCalculate |    .NET 4.7.2 |  1024 |  3,067.24 ns |    769.25 ns |    42.165 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|   SixLaborsCalculate |    .NET 4.7.2 |  1024 |  2,546.86 ns |  1,106.36 ns |    60.643 ns |  0.83 |    0.02 |     - |     - |     - |         - |
|                      |               |       |              |              |              |       |         |       |       |       |           |
| SharpZipLibCalculate | .NET Core 2.1 |  1024 |  3,377.15 ns |  3,903.41 ns |   213.959 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|   SixLaborsCalculate | .NET Core 2.1 |  1024 |  2,524.25 ns |  2,220.97 ns |   121.739 ns |  0.75 |    0.04 |     - |     - |     - |         - |
|                      |               |       |              |              |              |       |         |       |       |       |           |
| SharpZipLibCalculate | .NET Core 3.1 |  1024 |  3,980.60 ns |  8,497.37 ns |   465.769 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|   SixLaborsCalculate | .NET Core 3.1 |  1024 |     78.68 ns |     69.82 ns |     3.827 ns |  0.02 |    0.00 |     - |     - |     - |         - |
|                      |               |       |              |              |              |       |         |       |       |       |           |
| SharpZipLibCalculate |    .NET 4.7.2 |  2048 |  7,934.29 ns | 42,550.13 ns | 2,332.316 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|   SixLaborsCalculate |    .NET 4.7.2 |  2048 |  5,437.81 ns | 12,760.51 ns |   699.447 ns |  0.71 |    0.10 |     - |     - |     - |         - |
|                      |               |       |              |              |              |       |         |       |       |       |           |
| SharpZipLibCalculate | .NET Core 2.1 |  2048 |  6,008.05 ns |    621.37 ns |    34.059 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|   SixLaborsCalculate | .NET Core 2.1 |  2048 |  4,791.50 ns |  3,894.94 ns |   213.495 ns |  0.80 |    0.04 |     - |     - |     - |         - |
|                      |               |       |              |              |              |       |         |       |       |       |           |
| SharpZipLibCalculate | .NET Core 3.1 |  2048 |  5,900.06 ns |  1,344.70 ns |    73.707 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|   SixLaborsCalculate | .NET Core 3.1 |  2048 |    103.12 ns |     15.66 ns |     0.859 ns |  0.02 |    0.00 |     - |     - |     - |         - |
|                      |               |       |              |              |              |       |         |       |       |       |           |
| SharpZipLibCalculate |    .NET 4.7.2 |  4096 | 12,422.59 ns |  1,308.01 ns |    71.696 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|   SixLaborsCalculate |    .NET 4.7.2 |  4096 | 10,524.63 ns |  6,267.56 ns |   343.546 ns |  0.85 |    0.03 |     - |     - |     - |         - |
|                      |               |       |              |              |              |       |         |       |       |       |           |
| SharpZipLibCalculate | .NET Core 2.1 |  4096 | 11,888.00 ns |  1,059.25 ns |    58.061 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|   SixLaborsCalculate | .NET Core 2.1 |  4096 |  9,806.24 ns |    241.91 ns |    13.260 ns |  0.82 |    0.00 |     - |     - |     - |         - |
|                      |               |       |              |              |              |       |         |       |       |       |           |
| SharpZipLibCalculate | .NET Core 3.1 |  4096 | 12,181.28 ns |  1,974.68 ns |   108.239 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|   SixLaborsCalculate | .NET Core 3.1 |  4096 |    192.39 ns |     10.27 ns |     0.563 ns |  0.02 |    0.00 |     - |     - |     - |         - |

<!-- Thanks for contributing to ImageSharp! -->